+++ packages/strapi-admin/index.js
@@ -1,9 +1,11 @@
+/* eslint-disable no-useless-escape */
 const path = require('path');
 const fs = require('fs-extra');
 const webpack = require('webpack');
 const getWebpackConfig = require('./webpack.config.js');
 const WebpackDevServer = require('webpack-dev-server');
 const chalk = require('chalk');
+const chokidar = require('chokidar');
 
 const getPkgPath = name =>
   path.dirname(require.resolve(`${name}/package.json`));
@@ -137,6 +139,9 @@ async function createCacheDir(dir) {
 }
 
 async function build({ dir, env, options }) {
+  // Create the cache dir containing the front-end files.
+  await createCacheDir(dir);
+
   const cacheDir = path.resolve(dir, '.cache');
   const entry = path.resolve(cacheDir, 'admin', 'src', 'app.js');
   const dest = path.resolve(dir, 'build');
@@ -176,7 +181,10 @@ async function build({ dir, env, options }) {
   });
 }
 
-async function watch({ dir, port, options }) {
+async function watchAdmin({ dir, port, options }) {
+  // Create the cache dir containing the front-end files.
+  await createCacheDir(dir);
+
   const entry = path.join(dir, '.cache', 'admin', 'src', 'app.js');
   const dest = path.join(dir, 'build');
   const env = 'development';
@@ -193,9 +201,9 @@ async function watch({ dir, port, options }) {
     clientLogLevel: 'silent',
     hot: true,
     quiet: true,
-    publicPath: '/admin/',
+    publicPath: options.publicPath,
     historyApiFallback: {
-      index: '/admin/',
+      index: options.publicPath,
     },
   };
 
@@ -212,12 +220,108 @@ async function watch({ dir, port, options }) {
       chalk.green(`Admin development at http://localhost:${port}/admin`)
     );
   });
+
+  watchFiles(dir);
+}
+
+async function watchFiles(dir) {
+  const cacheDir = path.join(dir, '.cache');
+  const pkgJSON = require(path.join(dir, 'package.json'));
+  const admin = path.join(dir, 'admin');
+
+  const appPlugins = Object.keys(pkgJSON.dependencies).filter(
+    dep =>
+      dep.startsWith('strapi-plugin') &&
+      fs.existsSync(path.resolve(getPkgPath(dep), 'admin', 'src', 'index.js'))
+  );
+  const pluginsToWatch = appPlugins.map(plugin =>
+    path.join(
+      dir,
+      'extensions',
+      plugin.replace(/^strapi-plugin-/i, ''),
+      'admin'
+    )
+  );
+  const filesToWatch = [admin, ...pluginsToWatch];
+
+  const watcher = chokidar.watch(filesToWatch, {
+    ignoreInitial: true,
+    ignorePermissionErrors: true,
+  });
+
+  watcher.on('all', async (event, filePath) => {
+    const re = /\/extensions\/([^\/]*)\/.*$/gm;
+    const matched = re.exec(filePath);
+    const isExtension = matched !== null;
+    const pluginName = isExtension ? matched[1] : '';
+
+    const packageName = isExtension
+      ? `strapi-plugin-${pluginName}`
+      : 'strapi-admin';
+    const targetPath = isExtension
+      ? filePath.split('/extensions/')[1].replace(pluginName, '')
+      : filePath.split('/admin')[1];
+
+    const destFolder = isExtension
+      ? path.join(cacheDir, 'plugins', packageName)
+      : path.join(cacheDir, 'admin');
+
+    if (event === 'unlink' || event === 'unlinkDir') {
+      const originalFilePathInNodeModules = path.join(
+        getPkgPath(packageName),
+        isExtension ? '' : 'admin',
+        targetPath
+      );
+
+      // Remove the file or folder
+      // We need to copy the original files when deleting an override one
+      try {
+        fs.removeSync(path.join(destFolder, targetPath));
+      } catch (err) {
+        console.log('An error occured while deleting the file', err);
+      }
+
+      // Check if the file or folder exists in node_modules
+      // If so copy the old one
+      if (fs.pathExistsSync(path.resolve(originalFilePathInNodeModules))) {
+        try {
+          await fs.copy(
+            path.resolve(originalFilePathInNodeModules),
+            path.join(destFolder, targetPath)
+          );
+
+          // The plugins.js file needs to be recreated
+          // when we delete either the admin folder
+          // the admin/src folder
+          // or the plugins.js file
+          // since the path are different when developing inside the monorepository or inside an app
+          const shouldCopyPluginsJSFile =
+            filePath.split('/admin/src').filter(p => !!p).length === 1;
+
+          if (
+            (event === 'unlinkDir' &&
+              !isExtension &&
+              shouldCopyPluginsJSFile) ||
+            (!isExtension && filePath.includes('plugins.js'))
+          ) {
+            await createPluginsJs(appPlugins, path.join(cacheDir));
+          }
+        } catch (err) {
+          // Do nothing
+        }
+      }
+    } else {
+      // In any other case just copy the file into the .cache folder
+      try {
+        await fs.copy(filePath, path.join(destFolder, targetPath));
+      } catch (err) {
+        console.log(err);
+      }
+    }
+  });
 }
 
 module.exports = {
   build,
-  createPluginsJs,
-  createCacheDir,
-  watch,
-  copyCustomAdmin,
+  watchAdmin,
 };
