+++ packages/strapi-admin/package.json
@@ -102,6 +102,7 @@
   "license": "MIT",
   "gitHead": "c85658a19b8fef0f3164c19693a45db305dc07a9",
   "devDependencies": {
+    "chokidar": "^3.1.1",
     "webpack": "^4.40.1",
     "webpack-cli": "^3.3.2",
     "webpack-dev-server": "^3.4.1"
